+++ R/nls.R
@@ -30,10 +30,10 @@
 #'
 #' # Uses nlsLM from minpack.lm if available
 #'
-#' fit1 <- nlmixr(one.cmt, d, est="nls", nlsControl(algorithm="LM"))
+#' fit1 <- nlmixr(one.cmt, nlmixr2data::theo_sd, est="nls", nlsControl(algorithm="LM"))
 #'
 #' # Uses port and respect parameter boundaries
-#' fit2 <- nlmixr(one.cmt, d, est="nls", nlsControl(algorithm="port"))
+#' fit2 <- nlmixr(one.cmt, nlmixr2data::theo_sd, est="nls", nlsControl(algorithm="port"))
 #'
 #' # You can access the underlying nls object with `$nls`
 #' fit2$nls
@@ -656,7 +656,7 @@ rxUiGet.nlsFormula <- function(x, ...) {
   #.ret$etaObf <- NULL
   #.ret$omega <- NULL
   .ret$control <- .control
-  .ret$extra <- ""
+  .ret$extra <- paste0(" with ", crayon::bold$yellow(.control$algorithm),  " algorithm")
   .nlmixr2FitUpdateParams(.ret)
   nmObjHandleControlObject(.ret$control, .ret)
   if (exists("control", .ui)) {